requirements.txt
python-dotenv>=1.0.1
llama-cloud>=0.1.20
tqdm>=4.66.4
rapidfuzz>=3.9.6
sentence-transformers>=3.0.1
numpy>=1.26.4
neo4j>=5.23.1
pypdf>=4.3.1
